+++ manul/benches/empty_rounds.rs
@@ -9,9 +9,9 @@ use core::fmt::Debug;
 use criterion::{criterion_group, criterion_main, Criterion};
 use manul::{
     protocol::{
-        Artifact, Deserializer, DirectMessage, EchoBroadcast, EntryPoint, FinalizeError, FinalizeOutcome, LocalError,
-        NormalBroadcast, PartyId, Payload, Protocol, ProtocolError, ProtocolMessagePart, ProtocolValidationError,
-        ReceiveError, Round, RoundId, Serializer,
+        Artifact, BoxedRound, Deserializer, DirectMessage, EchoBroadcast, EntryPoint, FinalizeError, FinalizeOutcome,
+        LocalError, NormalBroadcast, PartyId, Payload, Protocol, ProtocolError, ProtocolMessagePart,
+        ProtocolValidationError, ReceiveError, Round, RoundId, Serializer,
     },
     session::{signature::Keypair, SessionOutcome},
     testing::{run_sync, BinaryFormat, TestSessionParams, TestSigner, TestVerifier},
@@ -75,16 +75,17 @@ struct Round1Artifact;
 
 impl<Id: PartyId> EntryPoint<Id> for EmptyRound<Id> {
     type Inputs = Inputs<Id>;
+    type Protocol = EmptyProtocol;
     fn new(
         _rng: &mut impl CryptoRngCore,
         _shared_randomness: &[u8],
         _id: Id,
         inputs: Self::Inputs,
-    ) -> Result<Self, LocalError> {
-        Ok(Self {
+    ) -> Result<BoxedRound<Id, Self::Protocol>, LocalError> {
+        Ok(BoxedRound::new_dynamic(Self {
             round_counter: 1,
             inputs,
-        })
+        }))
     }
 }
 
@@ -165,11 +166,11 @@ impl<Id: PartyId> Round<Id> for EmptyRound<Id> {
         if self.round_counter == self.inputs.rounds_num {
             Ok(FinalizeOutcome::Result(()))
         } else {
-            let round = EmptyRound {
+            let round = BoxedRound::new_dynamic(EmptyRound {
                 round_counter: self.round_counter + 1,
                 inputs: self.inputs,
-            };
-            Ok(FinalizeOutcome::another_round(round))
+            });
+            Ok(FinalizeOutcome::AnotherRound(round))
         }
     }