+++ manul/src/protocol/object_safe.rs
@@ -197,23 +197,44 @@ where
     }
 }
 
-// When we are wrapping types implementing Round and overriding `finalize()`,
-// we need to unbox the result of `finalize()`, set it as an attribute of the wrapping round,
-// and then box the result.
-//
-// Because of Rust's peculiarities, Box<dyn Round> that we return in `finalize()`
-// cannot be unboxed into an object of a concrete type with `downcast()`,
-// so we have to provide this workaround.
-impl<Id, P> dyn ObjectSafeRound<Id, Protocol = P>
-where
-    Id: PartyId,
-    P: Protocol,
-{
-    pub fn try_downcast<T: Round<Id>>(self: Box<Self>) -> Result<T, Box<Self>> {
-        if core::any::TypeId::of::<ObjectSafeRoundWrapper<Id, T>>() == self.get_type_id() {
-            // This should be safe since we just checked that we are casting to a correct type.
+// We do not want to expose `ObjectSafeRound` to the user, so it is hidden in a struct.
+/// A wrapped new round that may be returned by [`Round::finalize`]
+/// or [`EntryPoint::new`](`crate::protocol::EntryPoint::new`).
+#[derive_where::derive_where(Debug)]
+pub struct BoxedRound<Id: PartyId, P: Protocol> {
+    wrapped: bool,
+    round: Box<dyn ObjectSafeRound<Id, Protocol = P>>,
+}
+
+impl<Id: PartyId, P: Protocol> BoxedRound<Id, P> {
+    /// Wraps an object implementing the dynamic round trait ([`Round`](`crate::protocol::Round`)).
+    pub fn new_dynamic<R: Round<Id, Protocol = P>>(round: R) -> Self {
+        Self {
+            wrapped: true,
+            round: Box::new(ObjectSafeRoundWrapper::new(round)),
+        }
+    }
+
+    pub(crate) fn as_ref(&self) -> &dyn ObjectSafeRound<Id, Protocol = P> {
+        self.round.as_ref()
+    }
+
+    pub(crate) fn into_boxed(self) -> Box<dyn ObjectSafeRound<Id, Protocol = P>> {
+        self.round
+    }
+
+    fn boxed_type_is<T: 'static>(&self) -> bool {
+        core::any::TypeId::of::<T>() == self.round.get_type_id()
+    }
+
+    /// Attempts to extract an object of a concrete type, preserving the original on failure.
+    pub fn try_downcast<T: Round<Id>>(self) -> Result<T, Self> {
+        if self.wrapped && self.boxed_type_is::<ObjectSafeRoundWrapper<Id, T>>() {
+            // Safety: This is safe since we just checked that we are casting to the correct type.
             let boxed_downcast = unsafe {
-                Box::<ObjectSafeRoundWrapper<Id, T>>::from_raw(Box::into_raw(self) as *mut ObjectSafeRoundWrapper<Id, T>)
+                Box::<ObjectSafeRoundWrapper<Id, T>>::from_raw(
+                    Box::into_raw(self.round) as *mut ObjectSafeRoundWrapper<Id, T>
+                )
             };
             Ok(boxed_downcast.round)
         } else {
@@ -221,8 +242,27 @@ where
         }
     }
 
-    pub fn downcast<T: Round<Id>>(self: Box<Self>) -> Result<T, LocalError> {
+    /// Attempts to extract an object of a concrete type.
+    ///
+    /// Fails if the wrapped type is not `T`.
+    pub fn downcast<T: Round<Id>>(self) -> Result<T, LocalError> {
         self.try_downcast()
             .map_err(|_| LocalError::new(format!("Failed to downcast into type {}", core::any::type_name::<T>())))
+    }
+
+    /// Attempts to provide a reference to an object of a concrete type.
+    ///
+    /// Fails if the wrapped type is not `T`.
+    pub fn downcast_ref<T: Round<Id>>(&self) -> Result<&T, LocalError> {
+        if self.wrapped && self.boxed_type_is::<ObjectSafeRoundWrapper<Id, T>>() {
+            let ptr: *const dyn ObjectSafeRound<Id, Protocol = P> = self.round.as_ref();
+            // Safety: This is safe since we just checked that we are casting to the correct type.
+            Ok(unsafe { &*(ptr as *const T) })
+        } else {
+            Err(LocalError::new(format!(
+                "Failed to downcast into type {}",
+                core::any::type_name::<T>()
+            )))
+        }
     }
 }